requirements.txt
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
console==0.9907
cryptography==41.0.7
deprecation==2.1.0
ezenv==0.92
gotrue==2.1.0
h11==0.14.0
hexdump==3.3
httpcore==0.17.3
httpx==0.24.1
idna==3.6
Jinja2==3.1.2
jinxed==1.2.1
keyboard==0.13.5
MarkupSafe==2.1.3
nodeenv==1.8.0
numpy==1.26.2
packaging==23.2
postgrest==0.13.0
prisma==0.11.0
pyarrow==14.0.1
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pylance==0.8.17
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.2
requests==2.31.0
rsa==4.9
scapy==2.5.0
scapy-http==1.8.2
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.1
tomlkit==0.12.3
typing_extensions==4.8.0
urllib3==2.1.0
websocket-client==1.7.0
websockets==11.0.3
